リファクタリング: パッケージ構造の改善とモジュール分割
