+++ app/__init__.py
@@ -1,0 +1,5 @@
+"""
+Valorant Discord Bot for managing accounts and retrieving rank information
+"""
+
+__version__ = "1.0.0"
