+++ valodb.egg-info/SOURCES.txt
@@ -1,0 +1,17 @@
+README.md
+setup.py
+app/__init__.py
+app/accounts.py
+app/commands.py
+app/kabaneri.py
+app/keep_alive.py
+app/main.py
+app/modals.py
+app/spreadsheet.py
+app/valorant_api.py
+valodb.egg-info/PKG-INFO
+valodb.egg-info/SOURCES.txt
+valodb.egg-info/dependency_links.txt
+valodb.egg-info/entry_points.txt
+valodb.egg-info/requires.txt
+valodb.egg-info/top_level.txt
