+++ valodb.egg-info/requires.txt
@@ -4,5 +4,7 @@ valo_api==2.1.0
 gspread==5.12.1
 oauth2client==4.1.3
 PyNaCl==1.5.0
-setuptools==69.0.3
 python-dotenv==1.0.1
+
+[:python_version < "3.9"]
+zoneinfo
